Agregando categorias y productos desde la db

+++ src/main/java/com/avance/floreria/repository/CategoriaRepository.java
@@ -3,5 +3,5 @@ package com.avance.floreria.repository;
 import com.avance.floreria.entity.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 }
